diplomax_cm/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
httpx==0.27.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==43.0.1
pycryptodome==3.21.0
sqlalchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.3
redis==5.1.1
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0
requests==2.32.3
google-cloud-vision==3.8.1
firebase-admin==6.5.0
reportlab==4.2.5
Pillow==10.4.0
qrcode[pil]==7.4.2
boto3==1.35.36
python-dotenv==1.0.1
pytz==2024.2
aiofiles==24.1.0
celery[redis]==5.4.0
pytest==8.3.3
pytest-asyncio==0.24.0
